feat: Remove flutter_quill dependency from catalyst_voices_models, move to package document
